add tests, refactor tests, update cli notebook
