+++ update_api_docs.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+lucidoc eido -P rst > docs/api_docs.md
